donatebloodproj/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.5
django-crispy-forms==1.12.0
django-filter==2.4.0
django-tinymce==3.4.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.0.1
pytz==2021.1
requests==2.27.1
sqlparse==0.4.1
urllib3==1.26.9
